+++ src/app/api/axiosInstance.ts
@@ -1,5 +1,6 @@
 import useTokenStore from '@/store/useTokenStore';
 import axios from 'axios';
+import { refreshToken } from './auth';
 
 const headers = {
   'Content-Type': 'application/json',
@@ -10,6 +11,15 @@ const axiosInstance = axios.create({
   baseURL: process.env.NEXT_PUBLIC_BACKEND_URI,
   headers: headers,
   withCredentials: true,
+});
+
+axiosInstance.interceptors.request.use(async config => {
+  await refreshToken();
+  const token = useTokenStore.getState().accessToken;
+  if (token) {
+    config.headers['Authorization'] = useTokenStore.getState().accessToken;
+  }
+  return config;
 });
 
 axiosInstance.defaults.headers.common['Authorization'] =
